Speedup through improved requests and db commits
